Extended output
For the eye nothing changes, however the classes within Kitty were meant from the start to be extendable to multiple applications.
For the Kitty application nothing changes, but for the classes used a lot more can be done. I do for example have in mind to make it possible for the Kitty classes to be used in an app that will need to output everything in HTML... This will now be possible.

That's all! ;)